API/requirements.txt
matplotlib
numpy
pillow
tensorflow==2.5.0
fastapi
uvicorn
python-multipart
tensorflow-serving-api==2.5.0
